+++ .gitignore
@@ -1,0 +1,4 @@
+venv
+__pycache__
+.idea/**/workspace.xml
+.idea/**/tasks.xml
