+++ Analysis.py
@@ -1,0 +1,5 @@
+import streamlit as st
+
+
+def analysis():
+    st.info("Coming soon")
