+++ settings.py
@@ -1,0 +1,3 @@
+DOOR_IMG = "assets/door.png"
+DOOR_OPEN_IMG = "assets/open_door.png"
+DUCK_IMG = "assets/duck.jpeg"
